+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -180,7 +180,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2025-02-11T10:20:31.156Z",
+    "2025-02-11T11:00:13.258Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -214,7 +214,7 @@ export const FileListData$outboundSchema: z.ZodType<
   bytes: z.number(),
   fileName: z.string(),
   workspaceId: z.string(),
-  created: z.date().default(() => new Date("2025-02-11T10:20:31.156Z"))
+  created: z.date().default(() => new Date("2025-02-11T11:00:13.258Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {