+++ packages/orq-rc/RELEASES.md
@@ -389,3 +389,13 @@ Based on:
 - [typescript v3.1.0-rc.59] packages/orq-rc
 ### Releases
 - [NPM v3.1.0-rc.59] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.59 - packages/orq-rc
+
+## 2025-02-11 11:01:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.488.0 (2.506.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.1.0-rc.60] packages/orq-rc
+### Releases
+- [NPM v3.1.0-rc.60] https://www.npmjs.com/package/@orq-ai/node/v/3.1.0-rc.60 - packages/orq-rc